Fix import() not keeping event loop alive
Depending on the circumstances, an awaited import() could cause the
event loop to terminate prematurely if no other jobs were pending, as
module loading doesn't go through the Dispatch queue.  An additional
clause in the event loop exit condition is enough to fix it.

Also took the liberty of doing some more refactoring, renaming async.c
to dispatch.c in the process.