Fix spurious syntax errors in SSj
Unlike other commands, `eval` and `examine` use the entire command line,
verbatim, as the string to be evaluated; however, because the lexer is
generic it was still producing syntax errors during tokenization.  A
slight tweak to the tokenization strategy was enough to fix it: if a
syntax error is encountered, the remainder of the command line is used
as the final token.

Fixes #226.