Cause Game Crash on uncaught promise rejection
One of the most frustrating things about working with promise-based code
is that any exceptions thrown during execution of an async function or
promise reaction tend to get eaten.  With this change, unhandled promise
rejections are detected by the event loop and crash the game just like
synchronous exceptions do.

Special thanks goes to @rhuanjl for implementing the required
functionality in ChakraCore.