Add support for SPK-packaged system files
If a filename beginning with a `#/` exists in the SPK package, it will
be used in place of an asset by the same name in the physical system
directory.  This was the easy part.  It might be a bit more involved to
get Cell to package the system files in the first place...