Use weak references in JSAL when safe
When reading a property from an object stored on the JSAL stack, it's
safe to assume that if the stack reference is weak then the object is
otherwise reachable--and therefore the newly pushed values can be weak
references as well, avoiding the cost of a JsAddRef for every property
read.  This notably improves the performance of the VertexList
constructor.