Add Kami profiling module
Named after the Guardian of the Earth in Dragon Ball (and Z) and
Piccolo's counterpart, Kami.  Call `Kami.profile()` and pass an object
along with the name of a method to have that method automatically
monkey-patched with profiling code.  When your game exits the total time
running each method will be printed as well as the average time per
call.