Refactor SSj: source_t -> listing_t
The term "source" was used in a few different contexts in the SSj
codebase and IntelliSense also tends to get confused if there are
multiple structs with the same name, even when they don't conflict with
each other.  At least `listing_t` makes it clear the object refers only
to a source code listing and not, say, an actual source file.