Improve FlipScreen() behavior in Sphere v2 code
Now calling FlipScreen() will temporarily change the event loop exit
strategy to Sphere v1 mode until the next proper tick.  This allows
old-fashioned FlipScreen loops to be embedded in Sphere v2 code.

I've debated with myself for a long time whether or not to support this;
while it would have been a massive footgun when the engine was using
longjmp to skip the event loop, things are a lot more graceful now.