Remove useless `--no-sleep` command-line option
This option dates back to very early in miniSphere's life and has stuck
around for years despite never finding a use case.  Releasing the CPU
between frames ultimately didn't turn out to be a problem.  So, seeing
as the only purpose it seems to have is pointlessly heating up the CPU,
now's as good a time as any to retire it for good.