Add Surface#blendingMode property
This allows changing the blending mode for a surface, as can be done in
Sphere v1.  I'm not sure I'm entirely happy with this design--it
basically just mirrors the Sv1 setup--but mS 5.2 is still a few weeks
out so that gives me time to come up with a better design.  Implements
#257.