Rename a bunch of stuff for consistency
Avoid using "new" in the names of functions that return a struct
directly, since they don't need a corresponding free/unref.  Also
renamed colormatrix_t to color_fx_t to match with the applyColorFX
function.  The terminology used for this concept in the Sphere 1.x
documentation seems kind of inconsistent, to be honest...