Fix miniSphere segfault on SSj Blue detach
This is why I don't like maintaining two debugger clients.  It took me a
while to track the cause of this one down... turns out I renumbered the
SSj/Ki messages and forgot to update the Sphere Studio plugin to match.
This caused SSj Blue to send miniSphere a download request with no
arguments, ending in a segfault.