Harden Dispatch machinery against reentrancy
If `dispatch_run` happened to be reentered (due to a job calling
FlipScreen or MapEngine, for instance), the inner call would modify the
queue, often causing the outer call to segfault when it attempted to do
the same.  This fixes the bug by detecting when a reentrant call happens
and bailing out if so.