Add warning for .mjs main script in Sv1 build
For full backward compatibility, if the game manifest specifies Sphere
v1, the engine will run the main script as top-level program code
regardless of its extension.  Since `.mjs` is automatically treated as a
module in all other contexts, this behavior may be unexpected, so have
Cell produce a warning in that case.