Fix miniSphere crash with no CL args
When no game path was specified on the command line, `game_args_offset`
would never get initialized, causing the engine to misbehave once it
tried to pass command-line options to the game being launched.

To be honest, this is a massive showstopping bug and *never* should have
made it into a release.  If I hadn't skipped the release candidate
stage...