Reject games targeting experimental API
It occurred to me that allowing a game to target the current
experimental API level creates a forward-compatibility issue: games
developed in a later version of the engine and targeting the latest
stable API might end up requesting the experimental API in this version,
leading to potentially odd and difficult-to-debug behavior.  Not a good
idea.