Fix wrong mouse coordinates from Mouse#getEvent
Due to the input subsystem not correcting for window scaling and
letterboxing, the coordinates for Mouse#getEvent() would be incorrect if
either of those were active.  This mostly fixes it by applying the
correction to the returned event.  It's not perfect, as events generated
before a fullscreen/window switch will come out wrong, but should be
adequate in most real-world cases.