Excise all `jsal_push_eval` calls
In fact, `jsal_push_eval` itself no longer exists.  It was way too easy
to abuse, very expensive to call, and served no actual useful purpose in
the codebase (kind of like `eval` in JS, actually...) so getting rid of
it entirely will prevent falling into this trap in the future.