Further optimize Transform#matrix access
Since a function's magic number is now an `intptr_t` and a Transform's
underlying matrix pointer is stable, I can directly pass in a pointer to
the correct value rather than having to access the JS object every time.
Not only does this speed up normal access, it also streamlines the
initial setup of the accessors for new Transforms.