More Transform#matrix optimizations
This is about as fast as I can make it right now without changing the
observable behavior.  It's about 10 times faster for the first use of
.matrix compared to miniSphere 5.2.13, which is a massive improvement.

The improvement here largely comes from reusing a property descriptor
and an internal tweak to `jsal_def_prop_index` to avoid unnecessary
conversions and stack ops.