Optimize jsal_ref() to avoid JsAddRef
When taking a reference to a value which is strongly referenced by the
value stack (this is the usual case), we can avoid an addition pinning
operation by simply making the stack entry weak.  This avoids expensive
calls into ChakraCore but has the downside that it's not safe to release
the reference so long as the value remains on the stack.  This seems
like a safe assumption, given the cases where jsal_ref() is typically
used.