Use default fileTypes for system files
Fixes #281.  Without this fix, a game could easily break the whole
engine by associating a different file type with `.js`; this would
prevent the engine from being able to load the runtime modules, which
must be executed in ESM mode.  The fix is to only use the default type
mappings for anything in `#/`.