Implement recursive DirectoryStream in miniSphere
The code to support this was hastily adapted from what was already
written for Cell (this being my ulterior motive for switching miniSphere
to tinydir in the first place) but I'm not entirely happy with the
result.  It felt too much like shoving a square peg into a round hole
and there are a lot of inefficiencies as a result--notably an
unnecessary path_t/lstring_t roundtrip conversion that should be
refactored out as soon as possible.  That needs to happen quickly,
before I forgot about it...