Add note about paths in DirectoryStream#next()
For recursive streams, in lieu of returning actual directories,
`.fileName` can instead be a relative path.  This should be noted since
prior to the implementation of recursion, it was always safe to assume
`.fileName` didn't include any directory information.

I hasten to note that this is not a breaking change *per se* since
non-recursive streams (the default) will continue to behave as they
always have.