Fix buffer overflow in `strescq()` function
A combination of an off-by-one bug and a big oversight (not accounting
for the remainder of the string after all escapable characters are
counted) caused the buffer size for strescq() to be too small.  This
almost always leads to a segmentation fault, so I'm honestly surprised
it didn't show up in initial testing.