Implement Texture.fromFile() async API
Consider this the maiden voyage of #289.  Texture.fromFile() returns a
promise which resolves to the new texture object, rather than returning
the Texture object directly as with `new Texture`.  If Oozaru is going
to be a thing, the API will need to have functions like this in order to
be code-compatible with desktop miniSphere.