Implement first draft of FS.fetch()
FS.fetch() provides a generic way to load assets in both miniSphere and
Oozaru.  The plan is that it decides how to load a file automatically
based on its filename extension and `fileTypes` mapping, but that's not
implemented.  Right now it's largely an alias for require() but with
SphereFS path semantics.