Fix segfault on exit if ctor calls v1 functions
If the main module exports a class or non-async function, the engine
calls it as a constructor and then tries to call `.start()` on the
resulting instance.  If the constructor entered a FlipScreen or
MapEngine loop before returning however, closing the window would kill
the script by disabling the JS VM--leading to a segfault when the engine
tried to access the constructed object.

Easy enough to fix--just check that the VM is still enabled before
trying to call `.start()`.