Rename Sphere RT modules to '.js'
I kept putting this off because `require` will look in the system
directory for modules and potentially find these--then attempt to load
them as CommonJS and fail.  But really, `require` shouldn't even be
doing that in the first place now that ESM is the preferred module
system.  So I'll have to change that.