Add basic built-in support for from() queries
So far only where, select, reduce, and take are supported.  In order to
avoid the performance costs of jumping between the native and JS
domains, this uses an experimental method where the query is compiled
into a JS function so that once the query has started running,
everything happens on the JS side.