Implement from#ascending/descending() API
This provides a way to sort values in the middle of a query.  Normally
doing so would require running the query, but since we're generating JS
code anyway, we can just emit another `for` loop after sorting the
results.  Assuming the user *really* needs to do a sort in the middle of
a query, this should always be faster than running a second query.