Implement Query#find(), Query#toArray()
There's two more important functions I have yet to implement:
Query#any() and Query#all().  Once those are in I think I'll be able to
replace all uses of from() in Specs with the built-in function and then
this API addition will be complete!

...I do still need to refactoring the query compiler though.  It's kind
of a buffer overrun exploit waiting to happen.