Implement Query reverse, any, all, first, last
So far everything seems to be working flawlessly.  My crazy idea to
compile the query to JS was a rousing success and gives me a ton of
flexibility.  What it *doesn't* support yet is [Symbol.iterator] (and
therefore for..of).  That's going to have to change before this can make
it into a miniSphere release.