Implement Query 'anyIs' operator
Allows testing for a specific result value without the cost of calling a
function for every result.  This is similar in concept to
Array#includes(), but for queries.

I also took some time to refactor the handling for reducers.  There
ended up being a lot of different reductions and I had copy-and-pasted
code all over the place.