Make Query 'sample' and 'random' into QOPs
As in `from.js`, this allows continuing a query after taking a random
subset of values; sampling is often only the means and not an end in
itself.  Thanks to the optimization done for 'toArray', no speed is lost
if you actually *do* only need the sampled values.  In fact, the
implementation of 'sample' here is *faster* than what I had before!