Scrap built-in Query and query compiler
This was a fun little project, but given the number of bugs I've already
had to track down and fix, the performance gain is not worth the
additional maintenance cost.  Particularly not now that the included
'from.js' (which is much easier to maintain!) is so speedy itself.  It
actually has optimizations now that the built-in query engine
didn't--and would have been quite tricky to implement.