Disable relaxed sandbox in production
In the interest of getting #265 dealt with before miniSphere 5.3 is
released, this commit changes how the sandboxing mode in the game
manifest is handled.  It is now honored only by SpheRun (the first such
field of its kind!), and as a result the option has been moved into a
separate `development` namespace set aside for this selfsame purpose.

Fixes #265.