Disable loading ES module code using require()
`require` now always loads as either CommonJS or JSON, and never as ESM.
With this change, it's now possible to determine how a script is going
to be loaded entirely based on the code that loads it, without regard to
its file extension or any other non-obvious indicators.  `require` loads
CommonJS, `import` loads ESM and `FS.evaluateScript` loads traditional
JS--always.