Disable SphereFS sandbox for bare scripts
There's an argument to be made for being able to run bare scripts
without sandboxing, but I was hesitant to go that far by default as long
as it was still possible to run them in the end-user engine.  Now that
bare scripts only work in SpheRun, I feel more comfortable taking the
safety off.