Add internal support for API deprecation
Several API calls are slated for deprecation in API level 2, starting
with Surface#toTexture.  I don't like to write code before I actually
need it, but I've pretty well settled on how I want to go about this, so
may as well get the legwork out of the way now.