Replace `BlendOp` with new `Blender` class
This not only allows games to define custom blending modes, but since
any given blending operation is defined by up to 6 parameters, it avoids
repetition by not forcing the game to repeat all the parameters every
time it changes the blending mode.  Instead the game can create Blender
objects for commonly-used blend modes and reuse it as many times as it
needs.

The documentation will need to be updated before this makes it to
release.