Flesh out FileStream documentation
FileStream looks like it was the only built-in class without an
introductory paragraph explaining its purpose.  I figured now would be a
good time to add one, since I wanted to include a note that there's no
guarantee the file will be streamed directly from disk--since in the
case of Oozaru, no attempt is even made to do so.