Make Socket methods asynchronous
These should have been async from the start.  It's only an accident of
the way the `dyad.c` API is designed that synchronous reading from a
socket works at all and I don't ever expect that kind of thing to work
with, say, WebSockets.  As far as I know nobody is using the Socket API,
so I can afford a breaking change here in the name of portability.