Describe purpose of FileStream better
I'm still iffy on the whole "this can load the entire file into memory"
thing--it almost seems like if there's going to be something more
granular than FS.readFile() in the API then it should be optimized not
to have to fetch the entire file upfront.

On that note, now that JSON.fromFile() is a thing, `FS.readFile` would
probably be better returning an arraybuffer.