Refactor internal async call handling... again
For some reason when I implemented the async call support, I apparently
forgot that JSAL's call handling included a structure holding
information about the function.  So since the actual C call is handled
the same for async vs. sync calls (by design), it just needs to check
whether the function is flagged as async and if so, wrap the return
value in a promise.  No need for a separate callback for asyncs.