Move event loop into its own source file
The event loop is going to have new responsibilities very soon that
aren't directly related to the API layer, so it needs to be moved into
its own module.  Plus `pegasus.c` is huge, and it wasn't very good for
maintainability to have such an important component buried over 1,300
lines down.  There was a time the event loop was an afterthought, but
that's definitely not true anymore.