Add P1...P4 default joystick objects
This replaces the Joypad module.  I was hesitant when I originally
designed the API to provide default joystick mappings given the wide
variance in joystick and gamepad designs, but I suspect most games were
going to go right to the Joypad module anyway, so it might as well be in
the core.