Fix Socket#close destroying the socket
Sockets are meant to be reusable by design, to make it easier to deal
with lost connections.  Due to what I can only imagine was an oversight,
Socket#close actually destroyed the internal socket object, making it
impossible to reuse it--as well as causing a segfault if `connectTo` was
called afterwards.