Reinstate 64-bit Windows builds
So it turns out that a lot of JavaScript code runs up to 2x faster when
ChakraCore is compiled for x64 compared to x86.  While it was nice to
have a sub-7MB installer with only the 32-bit build, I'm not willing to
give up that much performance on 64-bit machines.