Check for valid save ID in Cell build
Since a missing save ID will prevent the use of the '~/' prefix--the
only place the SphereFS sandbox will allow a game to write data!--I
thought it would be nice if Cell warned the developer about it.  Also
added a check that the saveID only consists of alphanumeric characters
(and dots and dashes) to ensure it's always safe to use it as a
directory name.