Make `FS` functions synchronous again
As much as I really want to here, I can't go through with a breaking API
change, even if it's contingent on the targeted API level.  Plus if a
given platform supports fast synchronous I/O, the API really should
provide a way to use it, anyway.  I'm just going to have to suck it up
and add new async variants...