Add emptyPromises development option
Enabling emptyPromises disables fatal uncaught promise rejections; these
will be logged to the terminal but won't crash the game.  This is a
developer-only option, at least for now, as a rejected promise indicates
an error in some asynchronous operation that shouldn't be ignored in
production.  It's just that given the nature of asynchronous code in
general, and the fact that the error isn't caught until the end of the
frame (by design), the induced crash isn't always conducive to
debugging...