Restore 'Sphere.Game' object mutability
This was an unnecessary breaking change; it's enough just to document
that the engine won't act on any changes made to the object at runtime.
Plus Object.freeze() is shallow, so it was still possible to mutate
sub-objects.  Implementing a proper deep freeze would be more trouble
than it's worth.