miniSphere 5.3.0 RC1
API level 2 freeze, many new APIs, screenshot confirmation, improved
Cell command-line syntax, better sandboxing... and the list goes on and
on.  This is a HUGE release and the culmination of months of hard work.
Here's hoping I didn't break anything in the process!

Going to do a quick RC cycle for 5.3 since there were so many changes.
Let's see how long it takes to blow up...