Note that API functions can gain parameters
This is a potential hazard to code using higher-order functions, so it
should be noted.  Interestingly, of all the allowances made, this one
seems the most likely to produce breaking changes.  That said, it's
still better than ending up with 10 versions of the same function and
never being able to remove any of them...