Add version info to Windows installer
Just a small bit of polish.  If the user right-clicks the miniSphere
installer and clicks Properties, it will now say "miniSphere X.X.X for
Windows", where X.X.X is the version number.  This way even if the
installer is renamed it's always easy to see which engine version it's
for.