Improve SSj method performance in production
It occurred to me that, since the `SSj` functions don't actually do
anything in production, the no-ops can be implemented in JavaScript to
make that fact visible to the JIT and not only avoid a bunch of
expensive native calls, but also let the JS engine compile them away
completely when it realizes the functions being called are completely
empty.  This is especially important now that `SSj.assert` has been
added; nobody's going to use assertions if it means slowing down their
game.