Refactor `from.js`, rename some stuff
Added separate `size`, renamed `reduce` to `fold` (because the initial
value is required), and added internal helper functions for checking
whether something is array-like or iterable.

There are still too many loops; it would be nice to reuse the query
engine internally for operators like `over` and `plus`, but this tends
to cause a large performance hit so I don't want to do it.