Simplify semantics for Query#plus/without
Since this made it impossible to directly use any object with a `length`
property, I figured it was better if the operators don't try to be
clever.  I can add array-sourced versions later if the need proves
itself.

Technically a breaking change, but no code in the wild relies on the
original behavior to my knowledge, so this should be safe.