Disable write access to '$/' in Cell
While there are valid use cases for writing to the source tree in a
build script, it's easy for things to go horribly wrong so in the name
of SphereFS sandboxing, it should be disabled by default.

I'll eventually need to come up with an API that will allow Cellscripts
to re-enable write access to the source tree if they really need to.