Fix Cell SPK packager ignoring FS.writeFile()
The packager was only packing files associated with a target.  No target
was created internally for FS.writeFile(), only an entry in the list of
build artifacts.  This fixes it by having Cell pack all build artifacts,
which is technically a hack but will work for now because there is no
way to create a non-primary target.  More comprehensive refactoring will
happen for Cell 6.0.