Fold TTF support into main Font class
...and get rid of the extra `FontFace` class.  I'm very happy with how
this worked out.  It turned out I did think ahead when I designed the
`Font` class, which is the only thing that made this possible at all.  I
definitely want to do some internal refactoring though as layering the
TTF support on top of the existing RFN mechanisms ended up making
something of a mess.