Refactor module loader, consolidate some logic
This moves the entirety of the module resolution logic into its own
source file, `module.c`.  While I was at it, I refactored and rewrote a
huge chunk of the module resolver, improving compatibility with Node.js
and npm in the process.

Unfortunately, there are still two copies of the module resolver due to
differences between miniSphere's and Cell's file system interfaces.
Merging the two will need to wait until I'm able to generalize the
SphereFS code, which is easier said than done.