Remove node_modules from search paths again
I might add node_modules support to Cell, but it definitely has no place
in the Sphere implementation.  A few reasons for this: 1. There still
needs to be a Cell build step to put the modules into `dist` since the
module resolver can't search outside the SphereFS sandbox, in which case
they can just go into `lib` which is already designated for this
purpose, and 2. npm packages could never be supported in Oozaru anyway,
not even with import maps.