Fix issues with Xcode build
Building with Xcode failed due to differences in how it treats
angle-bracket includes (might be a compiler issue but it worked fine in
Linux); also `noreturn_t` is defined by one of miniSphere's
dependencies--I'm not sure which--which also caused the build to fail.
Both issues are fixed by this commit.