Fix some issues found by static analysis
Xcode isn't the greatest IDE, but the static analyzer it uses for C/C++
is kind of amazing.  This fixes a bunch of issues it found in SSj
including memory leaks, unused variables, and uninitialized variable
accesses.  I started with SSj because I figured it was the simplest;
miniSphere and Cell will likely be a bigger undertaking.