Show variable types with SSj `vars` command
Xcode's static analysis made me realize that I was retrieving the types
of variables when listing them but then didn't display the type info
anywhere.  Now it shows it; this is a big improvement over how it was
previously, where it just showed everything as `var`--which technically
wasn't even accurate now that `let` and `const` are supported!