SSj: Fix memory corruption in breakpoint handling
Caused by some recent refactoring.  A realloc() call was moved *before*
the associated array-length variable was incrementing, causing it to
allocate one less slot than needed and subsequently write past the end
of the array, leading to misbehavior.  This was fun to track down, since
it tended to only crash later, in unrelated code.  Alas, such is the
nature of heap corruption...