Repurpose Cell's `artifacts.json` file
Prior to this change, `artifacts.json was used only to track which files
Cell built previously, so that every single file didn't need to be
rebuilt each time.  It's still used for that, but it will also now be
used for general debugging information, so as not to bloat up
`game.json`.