Implement support for debugging with source maps
This gets miniSphere much closer to the goal of having first-class
TypeScript support.  Source maps and corresponding source code are
written by Cell to `artifacts.json`, allowing the debugger to easily
access them without having to breach the sandbox just to find source
code.  It's also easy to simply delete `artifacts.json` to remove all
the debugging information and convert the build to a release.