Rename the engine to "neoSphere"
This will be the official name of the desktop implementation going
forward.  It's larger than the original engine so it's not really "mini"
anymore, and since the full package has been renamed to Sphere, I don't
need to keep the miniSphere name for name recognition.  Now's as good a
time as any to go through with this, methinks.