Fix a few bugs in the `game.json` loader
All caused by the recent changes to treat `game.sgm` as the primary
source of truth for control information.  In particular, the the API
level was incorrectly being reset to 1 if it wasn't specified in the
JSON, even if it had already been set to something higher by the SGM
file.

I really need to clean up the `game.json` handling code.  Using the JSAL
API directly to do it wasn't one of my better ideas.