Reinstate Cellscript `describe` function
This function was removed back in miniSphere 4.6 in favor of letting the
Cellscript directly assign to `Sphere.Game`.  Unfortunately, that means
a lot of validation has to be postponed until the end of a build, when
the manifest(s) are written.  Having a dedicated function for this
should let validation be done at the time of the function call
instead--though note that this is not currently done.