Rebrand again, to neoSphere
Sphere is the name of the platform as a whole. Cell is reasonably
generic, but heavily neoSphere-centric, and SSj definitely won't work
with any other engine.  So since the repo is neoSphere-centric, that
makes the most sense for its name, I think.  Calling it just "Sphere"
implies that it's the entire platform in one repository, which... isn't
really the case.