Fix segfault on startup when no game is loaded
If no game is found on startup, `debugger_uninit()` will end up being
called without ever initializing the debugger.  Which was safe--except
for the sourcemap code, which wasn't prepared to deal with that
possibility due to an oversight on my part.

This might be a cue that I should clean up the initialization code...