SSj: Automatically step over Sphere Runtime code
This is likely not perfect, but serves as a proof-of-concept, at least.
If the debugger pauses within any module whose path starts with `#/`
(the system directory), the engine will keep stepping until the PC is
outside of the system code. Whether that means stepping in, out, or over
depends on what the previous stepping operation was.

I fully expect this first attempt to be buggy as it's a fairly naive
implementation, so more testing will be needed prior to the next
neoSphere release.