Remove background-loading APIs
Supporting `.whenReady()` added a ton of complexity just to support what
can already be done by simply holding onto the promise returned by a
`.fromFile()` call.  The thinking originally was that there were some
properties that could be accessed before the asset was fully loaded, but
it ended up not really being practical to support that.