Fix Xcode unreachable code warnings
Some map engine logic related to zones and triggers is intentionally
disabled for backward compatibility, but I wanted to leave the code in
in case I ever decided to enable it.  Unfortunately, Xcode kept nagging
me about unreachable code because of it.  Luckily, the warning was easy
to bypass by simply adding some parentheses.